Move controllers/pipelines LogKeys to an internal pkg
