+++ controllers/pipelines/workflow_repository.go
@@ -8,6 +8,7 @@ import (
 	pipelinesv1 "github.com/sky-uk/kfp-operator/apis/pipelines/v1alpha6"
 	"github.com/sky-uk/kfp-operator/controllers"
 	"k8s.io/apimachinery/pkg/labels"
+	"github.com/sky-uk/kfp-operator/controllers/pipelines/internal/logging"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/selection"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -68,7 +69,7 @@ func (w WorkflowRepositoryImpl) GetByLabels(ctx context.Context, matchingLabels 
 func (w WorkflowRepositoryImpl) MarkWorkflowAsProcessed(ctx context.Context, workflow *argo.Workflow) error {
 	logger := log.FromContext(ctx)
 
-	logger.V(1).Info("marking child workflow as processed", LogKeys.Workflow, workflow)
+	logger.V(1).Info("marking child workflow as processed", logging.LogKeys.Workflow, workflow)
 	workflowLabels := workflow.GetLabels()
 	if workflowLabels == nil {
 		workflowLabels = map[string]string{}
